[ Feature 🚀 ] Implemented Asymetric encryption algorithm RSA256 when signing and verifying JWT tokens.
Refactored code to add a new question instead of updating the current question.
