--- jwt-authentication/example.env
@@ -1,3 +1,0 @@
-ACCESS_TOKEN_SECRET=some_secret_key
-
-REFRESH_TOKEN_SECRET=some_secret_key
