+++ jwt-authentication/requests.rest
@@ -1,29 +1,25 @@
-GET http://localhost:3000/posts
-Authorization: Bearer eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyIjp7Im5hbWUiOiJhZG1pbiJ9LCJpYXQiOjE2NDI4NTA2NzUsImV4cCI6MTY0Mjg1NDI3NX0.ATn_q-pOlEfpsADdMslPIuRC1RK0i7eqEE9u_Ce9xv0iMyHBoYveMa-iS-IDfymrtfY1o5nhpil-8zoAtluRlVV3KdI4W5VOp0BnY16y3R-UaRW-pd8r6oLHvcCHKYXso3IgIIfan31BiFULIKit7btDoSJN-fqj2xk8LlAVDgBBSlqLGf1XbU8s0RvlhHWPxwZFKrJSfN3j5fud4jLUgpJoPAdwR0PxN9GYcISFRDzK0s_aFchvgdQWJ8QH0pbQGp759cVffB8eTgdFguse75QcIqsN53wllYa7m1S3oxmQgkCVy1GPkeaezUgEQ9w6xVILSiWVdEIHbOumG2nGS6WDGeJbsYuPHYe4ocR7jQiDbv3-0v9jiwGAuxuYqosTnc5pVBiW6U3U81MrG6v8ENcHYhz98LgMOFGBB4xckd1VP57vnAzQwV-lm2iiTHFYqt28G3-XZMtGQzAl_PCszCSekbS80nNjghD_7fsStEb5HBjRJiaqo-E6UIApdI1fxoSQkpjtsnE4EFzDCvETz2Up30uUVFr31DJFo8VsP046ZrlJz2Powgsa8Kj93lNk8sG2lEcge0FKrRG7SAsQ_857h1nCtNql6fdYm_yn1V6xBakZFeAchROih1mZixEpK4AzAKdIomc-4gn7vGCaHSzJlq8auLE2RK8XSm39uNA
+GET http://localhost:3000/posts Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJuYW1lIjoiSmltIiwiaWF0IjoxNTY4NzU5ODEyLCJleHAiOjE1Njg3NTk4Mjd9.0i6Im2gKz7jj8wM7aZZzOPaBS_xHoZWAqBwnldn-lQQ
 
 ###
 
-DELETE http://localhost:4000/logout
-Content-Type: application/json
+DELETE http://localhost:4000/logout Content-Type: application/json
 
 {
-  "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJuYW1lIjoiSmltIiwiaWF0IjoxNTY4NzU5OTIyfQ.RT6wszuCeFLwC_6ksmNMIELxiC5s-uRivfRxyZof5ag"
+"token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJuYW1lIjoiSmltIiwiaWF0IjoxNTY4NzU5OTIyfQ.RT6wszuCeFLwC_6ksmNMIELxiC5s-uRivfRxyZof5ag"
 }
 
 ###
 
-POST http://localhost:4000/token
-Content-Type: application/json
+POST http://localhost:4000/token Content-Type: application/json
 
 {
-  "token": "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyIjp7Im5hbWUiOiJhZG1pbiJ9LCJpYXQiOjE2NDI3OTY4MjksImV4cCI6MTY0Mjc5Njg4OX0.NaLCkDOGVT9vUzO7m367OmY85cG2LsSFnCzz05XGowPXqCLgxv1POwydN8M2ZoTJcSoW5OsGt7LBRaGbvuWfatKNjH6O8ToP-pnVI5USzs4cAapd9JaDpa57X5fCAmnk7FwBOVIs7tZzVcYle0w2MPOJNrODCkg6n8m3JKefYLDk9CLgE_gnI38dTPateA4lN6GliD3DiQm7oRe8KALbQD3710qe0cpW126Mim_7Sl2B2ltdpmWxvRa87V--LEtuejrhGiv1U9Vk-HbMlqv5mou_e7d2NVU3i1sJkAdJ6TTTzDShWSndzhRc2-e0usfKBID7yeEP5LzwGvDuQCYQfnmpzWMfPAyEBaobAoDv7Pzhsmo66L0P-86LBgYD8P0dO4Bjq-r2jgPC-Y2sV9rmj-g7AC9GfPgdxJmXu6s8gdu-dsamBtFZdstFsErSZ8221iW9KB_EZ8ya1WFXoDzgDWJo5cad5rjxADygPE4qUB6fQQucey4R07gksiK05kze4FLNFUKjRDXFxTmGgqnJvo2QgMzA3OcKWB8WJq6SvQmtMQMy7l3xRn6fyj_qMFIYpmP4vAIGW34YgWLh8XiW2ijnQNCoU31XE0OV_atPJRYHnNKh5BaFVrbFq2yorn3hSuOnBlFukitlJclLdNTgoQ3MqQDxivp9Jm8pERpnT5U"
+"token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJuYW1lIjoiSmltIiwiaWF0IjoxNTY4NzU5OTIyfQ.RT6wszuCeFLwC_6ksmNMIELxiC5s-uRivfRxyZof5ag"
 }
 
 ###
 
-POST http://localhost:4000/login
-Content-Type: application/json
+POST http://localhost:4000/login Content-Type: application/json
 
 {
-  "username": "admin"
+"username": "admin"
 }
