+++ sign-jwt-asymmetric-cryptography/example.env
@@ -1,0 +1,3 @@
+ACCESS_TOKEN_SECRET=some_secret_key
+
+REFRESH_TOKEN_SECRET=some_secret_key
